[MOV] web_editor: move image processing variables to utils
This commit moves some image processing variables to the utils file, in
order to be able to use them in a frontend bundle (specifically
`web_editor.assets_media_dialog`) in a further commit.

Related to task-3452267
